Fix building on Windows
More work needs to be done to test the app, but it is now possible to
build on Windows.

Some key changes include:

- Using the pre-built `onnxruntime` library
- Patching `esaxx-rs` to avoid statically linking crt
- Disabling telemetry on Windows builds at runtime

This overrides the previous `ort` branch, so the two upstream branches
still need to be unified, as it is not possible to have differing crate
sources per target.